requirements.txt
snowflake-connector-python
pyarrow>=10.0.1